fixup: Extracted FormattedLog into its own files
I left CustomLog.h alive because I suspect that CustomLog will gain some
more code in the foreseeable future and because that header helps
distinguish CustomLog-using code from code that builds other loggers on
top of FormattedLog (e.g., the upcoming Security::KeyLog).

Also polished CustomLog refactoring comment to allow for the (likely)
possibility of a class with an std::list data member.
