+++ src/log/FormattedLog.cc
@@ -12,8 +12,8 @@
 #include "cache_cf.h"
 #include "Debug.h"
 #include "log/Config.h"
-#include "log/CustomLog.h"
 #include "log/File.h"
+#include "log/FormattedLog.h"
 #include "Parsing.h"
 #include "sbuf/Stream.h"
 #include "SquidConfig.h"
